+++ app/src/main/java/com/indie/apps/pannypal/repository/UserRepository.kt
@@ -2,9 +2,9 @@ package com.indie.apps.pannypal.repository
 
 import com.indie.apps.pannypal.data.entity.User
 
-interface UserRepository : BaseRepository<User>{
+interface UserRepository : BaseRepository<User> {
 
-    suspend fun getUser() : User
+    suspend fun getUser(): User
 
-    suspend fun updateAmount(incomeAmt : Long, expenseAmt : Long): Int
+    suspend fun updateAmount(incomeAmt: Double, expenseAmt: Double): Int
 }
